Add testing for length of year
